update init to fix cases where the helper exits
When running the desktop-gnome-platform helper, it will exit part-way through if the platform interface isn't connected. Because the init part of the script is run before this then the needs_update/.last_revision will be set causing the helper to not compile the gnome schemas on the next launch, even if the platform interface is correctly connected. This causes gnome apps to die until either removed and reinstalled or an updated version is installed causing the .last_revision to no-longer match

this pr removes the setting of .last_revision in init deferring that to mark-and-exec.