desktop-exports: move user-dirs.* to $XDG_CONFIG_HOME
In case $HOME is not $SNAP_USER_DATA and XDG_CONFIG_HOME is not a subdir
of $HOME, move user-dirs.* file to the proper config folder.
Also trigger a xdg dirs local cached values reload (via update_xdg_dirs_values)
so that, if any of these folders changed, we can move files to the new position.