snapcraft.yaml: add audio-playback for audio
The pulseaudio interface is being deprecated in favor of audio-playback.
While the pulseaudio is still going to be available for a while, it will
stop being auto-connected for new snaps. Instead, developers should
plugs 'audio-playback' or both 'audio-playback' and 'audio-record' if
need to record audio. These new interfaces are available in snapd 2.41
and will be maintained for other media services, like pipewire.