suppress realpath noise on headless systems
When on systems where a desktop session has never run,
desktop-launch will create noise complaining that paths cannot be found.
Simply redirect the error messages to /dev/null.

E.G, running over SSH in Multipass, using snaps in a CLI environment.

Additionally, prevent behaviour where rmdir $HOME would be called
by ensuring the variable is not empty.
This is unlikely to have caused problems as rmdir would not delete dirs
that have content, but it's better to not rely on semantics here.